Version 1.0.10 Build 1 as submitted to the App Store.

+++ iSEPTAv2/Subclasses/GTFS/GTFSFunctions.m
@@ -82,7 +82,7 @@
 }
 
 
--(UInt32) getServiceIDFor:(NSString*) date
+-(NSInteger) getServiceIDFor:(NSString*) date
 {
     return 0;
 }
